pack stage struct to save storage/gas

+++ contracts/ERC721M.sol
@@ -198,10 +198,10 @@ contract ERC721M is IERC721M, ERC721AQueryable, Ownable {
 
     function updateStage(
         uint256 index,
-        uint256 price,
+        uint80 price,
         uint32 walletLimit,
         bytes32 merkleRoot,
-        uint256 maxStageSupply,
+        uint24 maxStageSupply,
         uint64 startTimeUnixSeconds,
         uint64 endTimeUnixSeconds
     ) external onlyOwner {
@@ -240,7 +240,7 @@ contract ERC721M is IERC721M, ERC721AQueryable, Ownable {
     function mint(
         uint32 qty,
         bytes32[] calldata proof,
-        uint256 timestamp,
+        uint64 timestamp,
         bytes calldata signature
     ) external payable {
         _mintInternal(qty, msg.sender, proof, timestamp, signature);
@@ -250,7 +250,7 @@ contract ERC721M is IERC721M, ERC721AQueryable, Ownable {
         uint32 qty,
         address to,
         bytes32[] calldata proof,
-        uint256 timestamp,
+        uint64 timestamp,
         bytes calldata signature
     ) external payable {
         if (_crossmintAddress == address(0)) revert CrossmintAddressNotSet();
@@ -265,7 +265,7 @@ contract ERC721M is IERC721M, ERC721AQueryable, Ownable {
         uint32 qty,
         address to,
         bytes32[] calldata proof,
-        uint256 timestamp,
+        uint64 timestamp,
         bytes calldata signature
     ) internal canMint hasSupply(qty) {
         if (_activeStage >= _mintStages.length) revert InvalidStage();
@@ -421,12 +421,12 @@ contract ERC721M is IERC721M, ERC721AQueryable, Ownable {
         revert InvalidStage();
     }
 
-    function _assertValidTimestamp(uint256 timestamp) internal view {
+    function _assertValidTimestamp(uint64 timestamp) internal view {
         if (timestamp < block.timestamp - MIN_STAGE_INTERVAL_SECONDS)
             revert TimestampExpired();
     }
 
-    function _assertValidStartAndEndTimestamp(uint256 start, uint256 end)
+    function _assertValidStartAndEndTimestamp(uint64 start, uint64 end)
         internal
         pure
     {
